renaming mui files into styles, new component folders.
