+++ src/components/OwnerCard/Owner.tsx
