+++ src/components/Questions/Questions.tsx
@@ -18,7 +18,7 @@ import { OnClick } from "../../types/eventType";
 
 //MUI components
 import CloudUploadIcon from "@material-ui/icons/CloudUpload";
-import { useStyles } from "./mui";
+import { useStyles } from "./styles";
 import SkipNextRoundedIcon from "@material-ui/icons/SkipNextRounded";
 import PlayCircleOutlineRoundedIcon from "@material-ui/icons/PlayCircleOutlineRounded";
