+++ src/components/RulesCard/styles.ts
@@ -1,0 +1,25 @@
+import { makeStyles, Theme, createStyles } from "@material-ui/core/styles";
+import { blue } from "@material-ui/core/colors";
+
+export const useStyles = makeStyles((theme: Theme) =>
+  createStyles({
+    root: {},
+    media: {
+      height: 0,
+      paddingTop: "56.25%", // 16:9
+    },
+    expand: {
+      transform: "rotate(0deg)",
+      marginLeft: "auto",
+      transition: theme.transitions.create("transform", {
+        duration: theme.transitions.duration.shortest,
+      }),
+    },
+    expandOpen: {
+      transform: "rotate(180deg)",
+    },
+    avatar: {
+      backgroundColor: blue[500],
+    },
+  })
+);
