+++ src/components/UserCard/mui.ts
@@ -1,0 +1,26 @@
+import { makeStyles, createStyles } from "@material-ui/core/styles";
+import { red } from "@material-ui/core/colors";
+
+export const useStyles = makeStyles((theme) =>
+  createStyles({
+    root: {},
+    media: {
+      paddingTop: "56.25%", // 16:9
+      alignItems: "center",
+      borderRadius: 10,
+    },
+    expand: {
+      transform: "rotate(0deg)",
+      marginLeft: "auto",
+      transition: theme.transitions.create("transform", {
+        duration: theme.transitions.duration.shortest,
+      }),
+    },
+    expandOpen: {
+      transform: "rotate(180deg)",
+    },
+    avatar: {
+      backgroundColor: red[500],
+    },
+  })
+);
