+++ src/components/navigation/NavBar/types.ts
@@ -1,0 +1,4 @@
+export type Props = {
+  darkMode: boolean;
+  set_darkMode: (value: boolean) => void;
+};
