+++ src/pages/Homepage/index.tsx
@@ -8,7 +8,7 @@ import QuestionCard from "../../components/Questions/Questions";
 import { useStyles } from "./mui";
 import Grid from "@material-ui/core/Grid";
 import { Paper } from "@material-ui/core";
-import Leaderboards from "../../components/TableCards/Leaderboards";
+import Leaderboards from "../LeaderBoards/Leaderboards";
 
 type Props = {
   darkMode: boolean;
