+++ src/pages/LeaderBoards/Leaderboards.tsx
@@ -1,17 +1,24 @@
 import React from "react";
 
+//Components
+import ScoreBoard from "../../components/Scoreboard";
+
 //Mui components
 import { makeStyles, Theme, createStyles } from "@material-ui/core/styles";
 import Card from "@material-ui/core/Card";
 import CardHeader from "@material-ui/core/CardHeader";
+
 import CardContent from "@material-ui/core/CardContent";
 import CardActions from "@material-ui/core/CardActions";
 import Collapse from "@material-ui/core/Collapse";
 import clsx from "clsx";
+
 import IconButton from "@material-ui/core/IconButton";
+
 import { blue } from "@material-ui/core/colors";
 import ExpandMoreIcon from "@material-ui/icons/ExpandMore";
-import { Box, CardMedia, Paper, TextField } from "@material-ui/core";
+
+import { Box, CardMedia, Paper } from "@material-ui/core";
 
 const useStyles = makeStyles((theme: Theme) =>
   createStyles({
@@ -36,7 +43,7 @@ const useStyles = makeStyles((theme: Theme) =>
   })
 );
 
-export default function RulezCard() {
+export default function Leaderboards() {
   const classes = useStyles();
   const [expanded, setExpanded] = React.useState(false);
 
@@ -46,11 +53,11 @@ export default function RulezCard() {
 
   return (
     <Paper elevation={23}>
-      <Card className={classes.root}>
-        <CardHeader title="My best results" subheader="result of all races" />
+      <Card>
+        <CardHeader title="Leaderboard" subheader="year race leaders" />
         <CardMedia
           className={classes.media}
-          image="https://cdn5.vectorstock.com/i/1000x1000/36/59/retro-pastel-color-smile-pink-brain-with-vector-10373659.jpg"
+          image="https://bojanzimmermann.com/wp-content/uploads/2018/03/Increase-Brain-Power-1024x677.png"
         />
 
         <CardActions disableSpacing>
@@ -68,14 +75,7 @@ export default function RulezCard() {
         <Collapse in={expanded} timeout="auto" unmountOnExit>
           <CardContent>
             <Box>
-              <TextField
-                id="outlined-multiline-static"
-                label="Multiline"
-                multiline
-                rows={4}
-                defaultValue="Default Value"
-                variant="outlined"
-              />
+              <ScoreBoard />
             </Box>
           </CardContent>
         </Collapse>
